Refactor http module into submodules and read db-type from header
